Refactor: move audio processing code into audio module (#597)
EQ/soft_mute/auto_negotiation/fallback/headroom/phase_alignment/audio_io を src/audio と include/audio に集約し、CMake と include 参照を更新してモジュール境界を明確化。
