+++ include/audio/auto_negotiation.h
