+++ src/audio/eq_parser.cpp
@@ -1,4 +1,4 @@
-#include "eq_parser.h"
+#include "audio/eq_parser.h"
 
 #include <algorithm>
 #include <cctype>
