+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "audio_io.h"
+#include "audio/audio_io.h"
 #include "convolution_engine.h"
 #include "core/config_loader.h"
 #include "core/filter_metadata.h"
