+++ tests/cpp/test_runtime_stats.cpp
@@ -1,5 +1,5 @@
+#include "audio/filter_headroom.h"
 #include "daemon/metrics/runtime_stats.h"
-#include "filter_headroom.h"
 #include "gtest/gtest.h"
 
 #include <atomic>
